+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Mediation/Vungle/Api/VungleInterstitialMediationExtras.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace GoogleMobileAds.Api.Mediation.Vungle
+namespace GoogleMobileAds.Mediation.Vungle.Api
 {
     public class VungleInterstitialMediationExtras : VungleMediationExtras
     {
@@ -11,5 +11,14 @@ namespace GoogleMobileAds.Api.Mediation.Vungle
         {
             get { return "com.google.unity.mediation.vungle.VungleUnityInterstitialExtrasBuilder"; }
         }
+    }
+}
+
+namespace GoogleMobileAds.Api.Mediation.Vungle
+{
+    [System.Obsolete("Use `GoogleMobileAds.Mediation.Vungle.Api.VungleInterstitialMediationExtras` instead.")]
+    public class VungleInterstitialMediationExtras :
+                 GoogleMobileAds.Mediation.Vungle.Api.VungleInterstitialMediationExtras
+    {
     }
 }