+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Mediation/Vungle/Api/VungleRewardedVideoMediationExtras.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace GoogleMobileAds.Api.Mediation.Vungle
+namespace GoogleMobileAds.Mediation.Vungle.Api
 {
     public class VungleRewardedVideoMediationExtras : VungleMediationExtras
     {
@@ -11,5 +11,15 @@ namespace GoogleMobileAds.Api.Mediation.Vungle
         {
             get { return "com.google.unity.mediation.vungle.VungleUnityRewardedVideoExtrasBuilder"; }
         }
+    }
+}
+
+
+namespace GoogleMobileAds.Api.Mediation.Vungle
+{
+    [System.Obsolete("Use `GoogleMobileAds.Mediation.Vungle.Api.VungleRewardedVideoMediationExtras` instead.")]
+    public class VungleRewardedVideoMediationExtras :
+                 GoogleMobileAds.Mediation.Vungle.Api.VungleRewardedVideoMediationExtras
+    {
     }
 }