+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Mediation/Vungle/Common/DummyClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Google LLC
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,9 +17,9 @@ using System.Reflection;
 
 using UnityEngine;
 
-using GoogleMobileAds.Api.Mediation.Vungle;
+using GoogleMobileAds.Mediation.Vungle.Api;
 
-namespace GoogleMobileAds.Common.Mediation.Vungle
+namespace GoogleMobileAds.Mediation.Vungle.Common
 {
     public class DummyClient : IVungleClient
     {